pyproject.toml
[tool.poetry]
name = "backend-test-monter"
version = "0.0.0"
description = "Meal delivery management."
authors = ["Jess Monter"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
djangorestframework = "^3.12.4"
django-redis = "^4.12.1"
gunicorn = "^20.1.0"
fluent-logger = "^0.10.0"
freezegun = "^1.1.0"
python-dateutil = "^2.8.1"
sentry-sdk = "^1.0.0"
psycopg2 = "^2.8.6"
pytz = "^2021.1"
pre-commit = "^2.12.1"
typed-ast = "^1.4.3"
redis = "^3.5.3"
celery = "^5.0.5"
python-dotenv = "^0.17.1"
django-rest-auth = "^0.9.5"
django-allauth = "^0.44.0"
djangorestframework-jwt = "^1.11.0"

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.3"
black = "^21.4b2"
isort = "^5.8.0"
pycodestyle = "^2.7.0"
pyflakes = "^2.3.1"
pylint = "^2.8.2"
pylint-django = "^2.4.4"
pylint-plugin-utils = "^0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
